+++ lib/Phpfastcache/Core/Pool/IO/IOHelperTrait.php
@@ -56,7 +56,7 @@ trait IOHelperTrait
             throw new PhpfastcacheIOException("Can't read PATH:" . $path);
         }
 
-        $stat->setData(implode(', ', array_keys($this->itemInstances)))
+        $stat->setData(implode(', ', \array_keys($this->itemInstances)))
             ->setRawData([
                 'tmp' => $this->tmp,
             ])
@@ -81,13 +81,13 @@ trait IOHelperTrait
         }
 
         $filename = $this->encodeFilename($keyword);
-        $folder = substr($filename, 0, 2) . DIRECTORY_SEPARATOR . substr($filename, 2, 2);
-        $path = rtrim($path, '/\\') . DIRECTORY_SEPARATOR . $folder;
+        $folder = \substr($filename, 0, 2) . DIRECTORY_SEPARATOR . \substr($filename, 2, 2);
+        $path = \rtrim($path, '/\\') . DIRECTORY_SEPARATOR . $folder;
 
         /**
          * Skip Create Sub Folders;
          */
-        if (!$skip && !is_dir($path) && @!mkdir($path, $this->getDefaultChmod(), true) && !is_dir($path)) {
+        if (!$skip && !\is_dir($path) && @!\mkdir($path, $this->getDefaultChmod(), true) && !\is_dir($path)) {
             throw new PhpfastcacheIOException('Path "' . $path . '" is not writable, please set a chmod 0777 or any writable permission and make sure to make use of an absolute path !');
         }
 
@@ -104,16 +104,16 @@ trait IOHelperTrait
         /**
          * Get the base system temporary directory
          */
-        $tmp_dir = rtrim(ini_get('upload_tmp_dir') ?: sys_get_temp_dir(), '\\/') . DIRECTORY_SEPARATOR . 'phpfastcache';
+        $tmp_dir = \rtrim(\ini_get('upload_tmp_dir') ?: \sys_get_temp_dir(), '\\/') . DIRECTORY_SEPARATOR . 'phpfastcache';
 
         /**
          * Calculate the security key
          */
         {
             $securityKey = $this->getConfig()->getSecurityKey();
-            if (!$securityKey || mb_strtolower($securityKey) === 'auto') {
+            if (!$securityKey || \mb_strtolower($securityKey) === 'auto') {
                 if (isset($_SERVER['HTTP_HOST'])) {
-                    $securityKey = preg_replace('/^www./', '', strtolower(str_replace(':', '_', $_SERVER['HTTP_HOST'])));
+                    $securityKey = \preg_replace('/^www./', '', \strtolower(\str_replace(':', '_', $_SERVER['HTTP_HOST'])));
                 } else {
                     $securityKey = ($this->isPHPModule() ? 'web' : 'cli');
                 }
@@ -130,12 +130,12 @@ trait IOHelperTrait
          * Extends the temporary directory
          * with the security key and the driver name
          */
-        $tmp_dir = rtrim($tmp_dir, '/') . DIRECTORY_SEPARATOR;
+        $tmp_dir = \rtrim($tmp_dir, '/') . DIRECTORY_SEPARATOR;
 
         if (empty($this->getConfig()->getPath())) {
             $path = $tmp_dir;
         } else {
-            $path = rtrim($this->getConfig()->getPath(), '/') . DIRECTORY_SEPARATOR;
+            $path = \rtrim($this->getConfig()->getPath(), '/') . DIRECTORY_SEPARATOR;
         }
 
         $path_suffix = $securityKey . DIRECTORY_SEPARATOR . $this->getDriverName();
@@ -150,27 +150,27 @@ trait IOHelperTrait
          * return the temp dir
          */
         if ($readonly === true) {
-            if ($this->getConfig()->isAutoTmpFallback() && (!@file_exists($full_path) || !@is_writable($full_path))) {
+            if ($this->getConfig()->isAutoTmpFallback() && (!@\file_exists($full_path) || !@\is_writable($full_path))) {
                 return $full_path_tmp;
             }
             return $full_path;
         }
 
-        if (!isset($this->tmp[$full_path_hash]) || (!@file_exists($full_path) || !@is_writable($full_path))) {
-            if (!@file_exists($full_path)) {
-                if (@mkdir($full_path, $this->getDefaultChmod(), true) === false && !is_dir($full_path)) {
+        if (!isset($this->tmp[$full_path_hash]) || (!@\file_exists($full_path) || !@\is_writable($full_path))) {
+            if (!@\file_exists($full_path)) {
+                if (@mkdir($full_path, $this->getDefaultChmod(), true) === false && !\is_dir($full_path)) {
                     throw new PhpfastcacheIOException('The directory ' . $full_path . ' could not be created.');
                 }
             } else {
-                if (!@is_writable($full_path)) {
-                    if (!@chmod($full_path, $this->getDefaultChmod()) && $this->getConfig()->isAutoTmpFallback()) {
+                if (!@\is_writable($full_path)) {
+                    if (!@\chmod($full_path, $this->getDefaultChmod()) && $this->getConfig()->isAutoTmpFallback()) {
                         /**
                          * Switch back to tmp dir
                          * again if the path is not writable
                          */
                         $full_path = $full_path_tmp;
-                        if (!@file_exists($full_path)) {
-                            if (@mkdir($full_path, $this->getDefaultChmod(), true) && !is_dir($full_path)) {
+                        if (!@\file_exists($full_path)) {
+                            if (@\mkdir($full_path, $this->getDefaultChmod(), true) && !\is_dir($full_path)) {
                                 throw new PhpfastcacheIOException('The directory ' . $full_path . ' could not be created.');
                             }
                         }
@@ -183,7 +183,7 @@ trait IOHelperTrait
              * writable including the temporary
              * one, we must throw an exception
              */
-            if (!@file_exists($full_path) || !@is_writable($full_path)) {
+            if (!@\file_exists($full_path) || !@\is_writable($full_path)) {
                 throw new PhpfastcacheIOException('Path "' . $full_path . '" is not writable, please set a chmod 0777 or any writable permission and make sure to make use of an absolute path !');
             }
 
@@ -207,7 +207,7 @@ trait IOHelperTrait
         ];
         $replace = ['-', '', ''];
 
-        return trim(preg_replace($regex, $replace, trim($filename)), '-');
+        return \trim(\preg_replace($regex, $replace, \trim($filename)), '-');
     }
 
     /**
@@ -230,9 +230,9 @@ trait IOHelperTrait
     protected function htaccessGen($path, $create = true)
     {
         if ($create === true) {
-            if (!is_writable($path)) {
+            if (!\is_writable($path)) {
                 try {
-                    if (!chmod($path, 0777)) {
+                    if (!\chmod($path, 0777)) {
                         throw new PhpfastcacheIOException('Chmod failed on : ' . $path);
                     }
                 } catch (PhpfastcacheIOException $e) {
@@ -240,8 +240,12 @@ trait IOHelperTrait
                 }
             }
 
-            if (!file_exists($path . "/.htaccess")) {
-                $content = <<<HTACCESS
+            if (!\file_exists($path . '/.htaccess')) {
+                $file = @\fopen($path . '/.htaccess', 'w+b');
+                if (!$file) {
+                    throw new PhpfastcacheIOException('PLEASE CHMOD ' . $path . ' - 0777 OR ANY WRITABLE PERMISSION!');
+                }
+                \fwrite($file, <<<HTACCESS
 ### This .htaccess is auto-generated by PhpFastCache ###
 <IfModule mod_authz_host>
 Require all denied
@@ -250,14 +254,8 @@ Require all denied
 Order Allow,Deny
 Deny from all
 </IfModule>
-HTACCESS;
-
-                $file = @fopen($path . '/.htaccess', 'w+b');
-                if (!$file) {
-                    throw new PhpfastcacheIOException('PLEASE CHMOD ' . $path . ' - 0777 OR ANY WRITABLE PERMISSION!');
-                }
-                fwrite($file, $content);
-                fclose($file);
+HTACCESS);
+                \fclose($file);
             }
         }
     }
@@ -278,21 +276,21 @@ HTACCESS;
      */
     protected function readFile($file): string
     {
-        if (function_exists('file_get_contents')) {
-            return (string)file_get_contents($file);
+        if (\function_exists('file_get_contents')) {
+            return (string)\file_get_contents($file);
         }
 
         $string = '';
 
-        $file_handle = @fopen($file, 'rb');
+        $file_handle = @\fopen($file, 'rb');
         if (!$file_handle) {
             throw new PhpfastcacheIOException("Cannot read file located at: {$file}");
         }
-        while (!feof($file_handle)) {
-            $line = fgets($file_handle);
+        while (!\feof($file_handle)) {
+            $line = \fgets($file_handle);
             $string .= $line;
         }
-        fclose($file_handle);
+        \fclose($file_handle);
 
         return $string;
     }
@@ -323,26 +321,26 @@ HTACCESS;
 
         if ($secureFileManipulation) {
             $tmpFilename = Directory::getAbsolutePath(dirname($file) . '/tmp_' . $this->getConfig()->getDefaultFileNameHashFunction()(
-                    str_shuffle(uniqid($this->getDriverName(), false))
-                    . str_shuffle(uniqid($this->getDriverName(), false))
+                    \str_shuffle(\uniqid($this->getDriverName(), false))
+                    . \str_shuffle(\uniqid($this->getDriverName(), false))
                 ));
 
-            $handle = fopen($tmpFilename, 'w+b');
-            if (is_resource($handle)) {
-                flock($handle, LOCK_EX);
+            $handle = \fopen($tmpFilename, 'w+b');
+            if (\is_resource($handle)) {
+                \flock($handle, \LOCK_EX);
                 $octetWritten = fwrite($handle, $data);
-                flock($handle, LOCK_UN);
-                fclose($handle);
+                \flock($handle, \LOCK_UN);
+                \fclose($handle);
             }
 
-            if (!rename($tmpFilename, $file)) {
-                throw new PhpfastcacheIOException(sprintf('Failed to rename %s to %s', $tmpFilename, $file));
+            if (!\rename($tmpFilename, $file)) {
+                throw new PhpfastcacheIOException(\sprintf('Failed to rename %s to %s', $tmpFilename, $file));
             }
         } else {
-            $handle = fopen($file, 'w+b');
-            if (is_resource($handle)) {
-                $octetWritten = fwrite($handle, $data);
-                fclose($handle);
+            $handle = \fopen($file, 'w+b');
+            if (\is_resource($handle)) {
+                $octetWritten = \fwrite($handle, $data);
+                \fclose($handle);
             }
         }